Sync official 10.4 release

+++ plumb.cpp
@@ -84,15 +84,15 @@ void plumb_update(const U32 curr_time_msec, const float getx, const float gety) 
    // Update player for diagnostic/table script visibility.  Only update if input value is larger than what's there.
    // When the table script reads the values, they will reset to 0. 
    if (TiltPerc > g_pplayer->m_ptable->m_tblNudgeReadTilt)
-	   g_pplayer->m_ptable->m_tblNudgeReadTilt = TiltPerc;
-   if (fabs(getx) > fabs(g_pplayer->m_ptable->m_tblNudgeReadX))
-	   g_pplayer->m_ptable->m_tblNudgeReadX = getx;
-   if (fabs(gety) > fabs(g_pplayer->m_ptable->m_tblNudgeReadY))
-	   g_pplayer->m_ptable->m_tblNudgeReadY = gety;
-   if (fabs(x) > fabs(g_pplayer->m_ptable->m_tblNudgePlumbX))
-	   g_pplayer->m_ptable->m_tblNudgePlumbX = x;
-   if (fabs(y) > fabs(g_pplayer->m_ptable->m_tblNudgePlumbY))
-	   g_pplayer->m_ptable->m_tblNudgePlumbY = y;
+      g_pplayer->m_ptable->m_tblNudgeReadTilt = TiltPerc;
+   if (fabsf(getx) > fabsf(g_pplayer->m_ptable->m_tblNudgeReadX))
+      g_pplayer->m_ptable->m_tblNudgeReadX = getx;
+   if (fabsf(gety) > fabsf(g_pplayer->m_ptable->m_tblNudgeReadY))
+      g_pplayer->m_ptable->m_tblNudgeReadY = gety;
+   if (fabsf(x) > fabsf(g_pplayer->m_ptable->m_tblNudgePlumbX))
+      g_pplayer->m_ptable->m_tblNudgePlumbX = x;
+   if (fabsf(y) > fabsf(g_pplayer->m_ptable->m_tblNudgePlumbY))
+      g_pplayer->m_ptable->m_tblNudgePlumbY = y;
 
    // Dampen the velocity.
    vx -= 2.50f * (vx * dt);
@@ -106,9 +106,9 @@ void plumb_update(const U32 curr_time_msec, const float getx, const float gety) 
       // This reduces annoying jittering when at rest.
       vx = 0.0f;
       vy = 0.0f;
-	  x = 0.0f;
-	  y = 0.0f;
-   } 
+      x = 0.0f;
+      y = 0.0f;
+   }
 
    // Update position.
    x += vx * dt;
